Fixes grunt/paths
I messed up the paths that `paths.src.all` pointed to with my previous
commit. I believe this fixes that issue. I'm not sure if `spec` or
`lib` are required at this point, but I know `grunt nag` and
`grunt make-standard` work as expected atm. Checking if `lib` and
`spec` are required now.